simultaneous code fixed
backendMergeRequest waits for users's click after entering time points
