+++ leaf-ui/js/loadSaveFunctions.js
@@ -11,7 +11,7 @@
 loader = document.getElementById("loader");
 layout_loader = document.getElementById("layout-loader");
 merge_button = document.getElementById("merge-button");
-merge_button_timing = document.getElementById("merge-button-timing");
+// merge_button_timing = document.getElementById("merge-button-timing");
 timing_input = document.getElementById("timing-input-window");
 merge_file_picker = document.getElementById("merge-file-picker");
 reader = new FileReader();
@@ -41,29 +41,30 @@ layout_loader.onchange = function () {
 };
 
 
-merge_button_timing.onclick = function(){
+// merge_button_timing.onclick = function(){
 	
-	let editedInputValues = [];
+	// let editedInputValues = [];
 
-	var intention_list = document.getElementById('timing-input-intention-list');
-	timeOrders = intention_list.getElementsByTagName("input");
+	// var intention_list = document.getElementById('timing-input-intention-list');
+	// timeOrders = intention_list.getElementsByTagName("input");
 	
-	// const { timing_list } = require('backendComm.js');
+	// // const { timing_list } = require('backendComm.js');
 
-	for (let i = 0; i < timeOrders.length; i++){
-		timeOrder = timeOrders[i].value;
-		editedInputValues.push(timeOrder);
-		console.log("edits: ", editedInputValues);
-		// timing_list[i].newTimeOrder = timeOrder;
-		globalTiming[i].newTimeOrder = timeOrder;
-		// console.log("globalTiming:", i," ", globalTiming[i].newTimeOrder);
-	}
-	timing_input.style.display = "none";
-	// backendMergeRequest(timing);
-	// console.log("new time Order: ", editedInputValues);
-	// console.log("globalTiming : ", globalTiming);
+	// for (let i = 0; i < timeOrders.length; i++){
+	// 	timeOrder = timeOrders[i].value;
+	// 	editedInputValues.push(timeOrder);
+	// 	// console.log("edits: ", editedInputValues);
+	// 	// timing_list[i].newTimeOrder = timeOrder;
+	// 	globalTiming[i].newTimeOrder = timeOrder;
+		
+	// 	// console.log("globalTiming:", i," ", globalTiming[i].newTimeOrder);
+	// }
+	// timing_input.style.display = "none";
+	// // backendMergeRequest(timing);
+	// // console.log("new time Order: ", editedInputValues);
+	// // console.log("globalTiming : ", globalTiming);
 
-}
+// }
 merge_button.onclick = function () {	
 	// console.log("Additional input")
 	var file1 = document.getElementById("merge-model1").files.item(0);
